Move dfx_generated up a dir for benchmarks
