+++ examples/motoko_examples/quicksort/dfx.json
@@ -8,7 +8,7 @@
             "candid": "canisters/azle/index.did",
             "wasm": "target/wasm32-unknown-unknown/release/azle.wasm",
             "declarations": {
-                "output": "test/dfx_generated/azle"
+                "output": "dfx_generated/azle"
             }
         }
     }
